+++ src/codegen/java/edu/ie3/codegen/MethodGenerator.java
@@ -5,7 +5,6 @@
 */
 package edu.ie3.codegen;
 
-import static edu.ie3.codegen.ResolverUtils.resolveClassName;
 import static edu.ie3.codegen.ResolverUtils.resolveType;
 
 import com.palantir.javapoet.*;
@@ -84,45 +83,36 @@ public class MethodGenerator implements HelperMethods {
     List<MethodSpec> methods = new ArrayList<>();
 
     for (GenerationConfig.MethodOverride override : genConfig.methodOverrides) {
-      var methodBuilder =
-          MethodSpec.methodBuilder(override.name)
-              .returns(resolveType(override.type))
-              .addAnnotation(Override.class)
-              .addModifiers(Modifier.PUBLIC);
+      override.annotation = true;
 
-      if (override.className != null && !override.className.isBlank()) {
-        methodBuilder.addStatement(
-            "return $T." + override.expression, resolveClassName(override.className));
+      var methodBuilder =
+          MethodSpec.methodBuilder(override.name).returns(resolveType(override.type));
+
+      if (override.isProtected) {
+        methodBuilder.addModifiers(Modifier.PROTECTED);
       } else {
-        methodBuilder.addStatement("return " + override.expression);
+        methodBuilder.addModifiers(Modifier.PUBLIC);
       }
 
-      methods.add(methodBuilder.build());
+      methods.add(enrichBuilder(methodBuilder, override).build());
     }
 
     for (GenerationConfig.MethodInsert insert : genConfig.methodInserts) {
-      var methodBuilder =
-          MethodSpec.methodBuilder(insert.name)
-              .returns(resolveType(insert.type))
-              .addModifiers(Modifier.PUBLIC);
+      var methodBuilder = MethodSpec.methodBuilder(insert.name).returns(resolveType(insert.type));
 
-      if (!insert.javaDoc.isBlank()) {
-        methodBuilder.addJavadoc(insert.javaDoc);
+      if (insert.isStatic) {
+        methodBuilder.addModifiers(Modifier.STATIC);
       }
 
-      if (insert.isAbstract) {
-        methodBuilder.addModifiers(Modifier.ABSTRACT);
-
+      if (insert.isPrivate && !insert.isProtected) {
+        methodBuilder.addModifiers(Modifier.PRIVATE);
+      } else if (insert.isProtected) {
+        methodBuilder.addModifiers(Modifier.PROTECTED);
       } else {
-        if (insert.className != null && !insert.className.isBlank()) {
-          methodBuilder.addStatement(
-              "return $T." + insert.expression, resolveClassName(insert.className));
-        } else {
-          methodBuilder.addStatement("return " + insert.expression);
-        }
+        methodBuilder.addModifiers(Modifier.PUBLIC);
       }
 
-      methods.add(methodBuilder.build());
+      methods.add(enrichBuilder(methodBuilder, insert).build());
     }
 
     if (genConfig.equals) {
@@ -149,11 +139,10 @@ public class MethodGenerator implements HelperMethods {
             .addParameter(Object.class, "o");
 
     builder.addStatement("if (this == o) return true");
-    builder.addStatement("if (!(o instanceof $L that)) return false", model.name);
 
     List<ModelDefinition.ComponentDefinition> filteredComponents = new ArrayList<>();
     for (ModelDefinition.ComponentDefinition component : model.components) {
-      if (excludeFromMethods(component)) {
+      if (excludeFromMethods(component, genConfig.excludeFromMethods)) {
         continue;
       }
       filteredComponents.add(component);
@@ -163,12 +152,14 @@ public class MethodGenerator implements HelperMethods {
 
     if (filteredComponents.isEmpty()) {
       if (superStatement) {
+        builder.addStatement("if (!(o instanceof $L that)) return false", model.name);
         builder.addStatement("return super.equals(o)");
       } else {
-        builder.addStatement("if (!super.equals(o)) return false");
-        builder.addStatement("return true");
+        builder.addStatement("return o instanceof $L that", model.name);
       }
     } else {
+      builder.addStatement("if (!(o instanceof $L that)) return false", model.name);
+
       if (superStatement) {
         builder.addStatement("if (!super.equals(o)) return false");
       }
@@ -186,7 +177,7 @@ public class MethodGenerator implements HelperMethods {
 
         if (isPrimitive(type)) {
           expression.add("$L == that.$L", component.name, component.name);
-        } else if (useEquals(type)) {
+        } else if (useEquals(type) && component.required) {
           expression.add("$L.equals(that.$L)", component.name, component.name);
         } else {
           expression.add("$T.equals($L, that.$L)", Objects.class, component.name, component.name);
@@ -213,7 +204,7 @@ public class MethodGenerator implements HelperMethods {
     }
 
     for (ModelDefinition.ComponentDefinition component : model.components) {
-      if (excludeFromMethods(component)) {
+      if (excludeFromMethods(component, genConfig.excludeFromMethods)) {
         continue;
       }
 
@@ -245,7 +236,7 @@ public class MethodGenerator implements HelperMethods {
 
     int index = 0;
     for (ModelDefinition.ComponentDefinition component : allComponents) {
-      if (excludeFromMethods(component)) {
+      if (excludeFromMethods(component, genConfig.excludeFromMethods)) {
         continue;
       }