+++ src/main/java/edu/ie3/datamodel/models/value/HeatDemandValue.java
@@ -1,22 +1,22 @@
 /*
- * © 2021. TU Dortmund University,
+ * © 2026. TU Dortmund University,
  * Institute of Energy Systems, Energy Efficiency and Energy Economics,
  * Research group Distribution grid planning and operation
 */
 package edu.ie3.datamodel.models.value;
 
-import edu.ie3.datamodel.models.StandardUnits;
 import java.util.Objects;
 import java.util.Optional;
 import javax.measure.quantity.Power;
 import tech.units.indriya.ComparableQuantity;
 
-/** Describes as heat demand value */
+/** Describes as heat demand value. */
 public class HeatDemandValue implements Value {
+  /** Heat demand as power. */
   private final ComparableQuantity<Power> heatDemand;
 
   public HeatDemandValue(ComparableQuantity<Power> heatDemand) {
-    this.heatDemand = heatDemand == null ? null : heatDemand.to(StandardUnits.HEAT_DEMAND_PROFILE);
+    this.heatDemand = heatDemand;
   }
 
   public Optional<ComparableQuantity<Power>> getHeatDemand() {
@@ -26,8 +26,7 @@ public class HeatDemandValue implements Value {
   @Override
   public boolean equals(Object o) {
     if (this == o) return true;
-    if (o == null || getClass() != o.getClass()) return false;
-    HeatDemandValue that = (HeatDemandValue) o;
+    if (!(o instanceof HeatDemandValue that)) return false;
     return Objects.equals(heatDemand, that.heatDemand);
   }
 
@@ -38,6 +37,6 @@ public class HeatDemandValue implements Value {
 
   @Override
   public String toString() {
-    return "HeatDemandValue{" + "heatDemand=" + heatDemand + '}';
+    return "HeatDemandValue{" + "heatDemand=" + heatDemand + "}";
   }
 }