+++ src/main/java/edu/ie3/datamodel/models/value/TemperatureValue.java
@@ -1,26 +1,25 @@
 /*
- * © 2021. TU Dortmund University,
+ * © 2026. TU Dortmund University,
  * Institute of Energy Systems, Energy Efficiency and Energy Economics,
  * Research group Distribution grid planning and operation
 */
 package edu.ie3.datamodel.models.value;
 
-import edu.ie3.datamodel.models.StandardUnits;
 import java.util.Objects;
 import java.util.Optional;
 import javax.measure.quantity.Temperature;
 import tech.units.indriya.ComparableQuantity;
 
-/** Describes a temperature value */
+/** Describes a temperature value. */
 public class TemperatureValue implements Value {
-  /** Temperature (typically in K) */
+  /** Temperature (typically in K). */
   private final ComparableQuantity<Temperature> temperature;
 
   /**
    * @param temperature (typically in K)
    */
   public TemperatureValue(ComparableQuantity<Temperature> temperature) {
-    this.temperature = temperature == null ? null : temperature.to(StandardUnits.TEMPERATURE);
+    this.temperature = temperature;
   }
 
   public Optional<ComparableQuantity<Temperature>> getTemperature() {
@@ -30,8 +29,7 @@ public class TemperatureValue implements Value {
   @Override
   public boolean equals(Object o) {
     if (this == o) return true;
-    if (o == null || getClass() != o.getClass()) return false;
-    TemperatureValue that = (TemperatureValue) o;
+    if (!(o instanceof TemperatureValue that)) return false;
     return Objects.equals(temperature, that.temperature);
   }
 
@@ -42,6 +40,6 @@ public class TemperatureValue implements Value {
 
   @Override
   public String toString() {
-    return "TemperatureValue{" + "temperature=" + temperature + '}';
+    return "TemperatureValue{" + "temperature=" + temperature + "}";
   }
 }