+++ src/main/java/edu/ie3/datamodel/models/value/VoltageValue.java
@@ -1,50 +1,40 @@
 /*
- * © 2024. TU Dortmund University,
+ * © 2026. TU Dortmund University,
  * Institute of Energy Systems, Energy Efficiency and Energy Economics,
  * Research group Distribution grid planning and operation
 */
 package edu.ie3.datamodel.models.value;
-
-import static edu.ie3.datamodel.models.StandardUnits.VOLTAGE_ANGLE;
-import static edu.ie3.util.quantities.PowerSystemUnits.DEGREE_GEOM;
-import static edu.ie3.util.quantities.PowerSystemUnits.PU;
-import static java.lang.Math.*;
 
 import java.util.Objects;
 import java.util.Optional;
 import javax.measure.quantity.Angle;
 import javax.measure.quantity.Dimensionless;
 import tech.units.indriya.ComparableQuantity;
-import tech.units.indriya.quantity.Quantities;
 
-/** Describes a voltage value as a pair of magnitude and angle */
+/** Describes a voltage value as a pair of magnitude and angle. */
 public class VoltageValue implements Value {
-
   /** Magnitude of the voltage in p.u. */
   private final ComparableQuantity<Dimensionless> magnitude;
 
-  /** Angle of the voltage in degree */
+  /** Angle of the voltage in degree. */
   private final ComparableQuantity<Angle> angle;
 
   /**
    * @param magnitude of the voltage in p.u.
-   * @param angleOption option for the angle of this voltage in degree
-   */
-  public VoltageValue(
-      ComparableQuantity<Dimensionless> magnitude,
-      Optional<ComparableQuantity<Angle>> angleOption) {
-    this.magnitude = magnitude;
-    this.angle = angleOption.orElse(Quantities.getQuantity(0.0, VOLTAGE_ANGLE));
-  }
-
-  /**
-   * @param magnitude of the voltage in p.u.
-   * @param angle of the voltage in degree
+   * @param angle of this voltage in degree
    */
   public VoltageValue(
       ComparableQuantity<Dimensionless> magnitude, ComparableQuantity<Angle> angle) {
     this.magnitude = magnitude;
     this.angle = angle;
+  }
+
+  /**
+   * @param magnitude of the voltage in p.u.
+   */
+  public VoltageValue(ComparableQuantity<Dimensionless> magnitude) {
+    this.magnitude = magnitude;
+    this.angle = null;
   }
 
   public Optional<ComparableQuantity<Dimensionless>> getMagnitude() {
@@ -55,27 +45,10 @@ public class VoltageValue implements Value {
     return Optional.ofNullable(angle);
   }
 
-  public Optional<ComparableQuantity<Dimensionless>> getRealPart() {
-    double mag = magnitude.to(PU).getValue().doubleValue();
-    double ang = angle.to(DEGREE_GEOM).getValue().doubleValue();
-
-    double eInPu = mag * cos(toRadians(ang));
-    return Optional.of(Quantities.getQuantity(eInPu, PU));
-  }
-
-  public Optional<ComparableQuantity<Dimensionless>> getImagPart() {
-    double mag = magnitude.to(PU).getValue().doubleValue();
-    double ang = angle.to(DEGREE_GEOM).getValue().doubleValue();
-
-    double fInPu = mag * sin(toRadians(ang));
-    return Optional.of(Quantities.getQuantity(fInPu, PU));
-  }
-
   @Override
   public boolean equals(Object o) {
     if (this == o) return true;
-    if (o == null || getClass() != o.getClass()) return false;
-    VoltageValue that = (VoltageValue) o;
+    if (!(o instanceof VoltageValue that)) return false;
     return Objects.equals(magnitude, that.magnitude) && Objects.equals(angle, that.angle);
   }
 
@@ -86,6 +59,6 @@ public class VoltageValue implements Value {
 
   @Override
   public String toString() {
-    return "VoltageValue{" + "magnitude=" + magnitude + ", angle=" + angle + '}';
+    return "VoltageValue{" + "magnitude=" + magnitude + ", angle=" + angle + "}";
   }
 }