+++ src/test/groovy/edu/ie3/test/common/IconWeatherTestData.groovy
@@ -24,8 +24,8 @@ class IconWeatherTestData extends WeatherTestData {
   Quantities.getQuantity(24.4741992659816, StandardUnits.TEMPERATURE),
   Quantities.getQuantity(270.45278309919627, StandardUnits.WIND_DIRECTION),
   Quantities.getQuantity(3.76601470961371, StandardUnits.WIND_VELOCITY),
-  Optional.of(Quantities.getQuantity(-5.0, StandardUnits.TEMPERATURE)),
-  Optional.of(Quantities.getQuantity(-4.0, StandardUnits.TEMPERATURE)),
+  Quantities.getQuantity(-5.0, StandardUnits.TEMPERATURE),
+  Quantities.getQuantity(-4.0, StandardUnits.TEMPERATURE),
   )
 
   public static final WeatherValue WEATHER_VALUE_67775_16H = new WeatherValue(
@@ -35,8 +35,8 @@ class IconWeatherTestData extends WeatherTestData {
   Quantities.getQuantity(24.1700023473353, StandardUnits.TEMPERATURE),
   Quantities.getQuantity(278.144331776102, StandardUnits.WIND_DIRECTION),
   Quantities.getQuantity(4.05744164637287, StandardUnits.WIND_VELOCITY),
-  Optional.of(Quantities.getQuantity(-3.0, StandardUnits.TEMPERATURE)),
-  Optional.of(Quantities.getQuantity(-2.0, StandardUnits.TEMPERATURE)),
+  Quantities.getQuantity(-3.0, StandardUnits.TEMPERATURE),
+  Quantities.getQuantity(-2.0, StandardUnits.TEMPERATURE),
   )
 
   public static final WeatherValue WEATHER_VALUE_67775_17H = new WeatherValue(
@@ -46,8 +46,8 @@ class IconWeatherTestData extends WeatherTestData {
   Quantities.getQuantity(23.6787403584074, StandardUnits.TEMPERATURE),
   Quantities.getQuantity(286.891007103442, StandardUnits.WIND_DIRECTION),
   Quantities.getQuantity(3.81526300455393, StandardUnits.WIND_VELOCITY),
-  Optional.of(Quantities.getQuantity(-1.0, StandardUnits.TEMPERATURE)),
-  Optional.of(Quantities.getQuantity(0.0, StandardUnits.TEMPERATURE)),
+  Quantities.getQuantity(-1.0, StandardUnits.TEMPERATURE),
+  Quantities.getQuantity(0.0, StandardUnits.TEMPERATURE),
   )
 
   public static final WeatherValue WEATHER_VALUE_67776_15H = new WeatherValue(
@@ -57,8 +57,8 @@ class IconWeatherTestData extends WeatherTestData {
   Quantities.getQuantity(22.365335568404, StandardUnits.TEMPERATURE),
   Quantities.getQuantity(245.604554131632, StandardUnits.WIND_DIRECTION),
   Quantities.getQuantity(4.39390441381814, StandardUnits.WIND_VELOCITY),
-  Optional.of(Quantities.getQuantity(1.0, StandardUnits.TEMPERATURE)),
-  Optional.of(Quantities.getQuantity(2.0, StandardUnits.TEMPERATURE)),
+  Quantities.getQuantity(1.0, StandardUnits.TEMPERATURE),
+  Quantities.getQuantity(2.0, StandardUnits.TEMPERATURE),
   )
 
   public static final WeatherValue WEATHER_VALUE_67776_16H = new WeatherValue(
@@ -68,7 +68,7 @@ class IconWeatherTestData extends WeatherTestData {
   Quantities.getQuantity(20.305111314491, StandardUnits.TEMPERATURE),
   Quantities.getQuantity(252.810224701109, StandardUnits.WIND_DIRECTION),
   Quantities.getQuantity(3.44242472583919, StandardUnits.WIND_VELOCITY),
-  Optional.of(Quantities.getQuantity(3.0, StandardUnits.TEMPERATURE)),
-  Optional.of(Quantities.getQuantity(4.0, StandardUnits.TEMPERATURE)),
+  Quantities.getQuantity(3.0, StandardUnits.TEMPERATURE),
+  Quantities.getQuantity(4.0, StandardUnits.TEMPERATURE),
   )
 }